+++ tests/test_rectangle.py
@@ -1,5 +1,5 @@
 import pytest
-from autoprotocol_utilities.rectangle import chop_list, binary_list, max_histogram_area, max_rectangle
+from autoprotocol_utilities.rectangle import chop_list, binary_list, max_histogram_area, max_rectangle, get_quadrant_well
 
 
 @pytest.mark.parametrize("wells, chop_length, r", [
@@ -51,3 +51,13 @@ def test_max_rectangle(wells, value, r):
     assert rect.height == r[1]
     assert rect.x == r[2]
     assert rect.y == r[3]
+
+
+@pytest.mark.parametrize("well, quad, r", [
+    ([0], 0, 0),
+    ([0], 1, 1),
+    ([0], 2, 24),
+    ([0], 3, 25)
+])
+def test_get_quadrant_well(well, quad, r):
+    assert(get_quadrant_well(well, quad)[0] == r)